requirements.txt
beautifulsoup4==4.11.1
Django==4.1.5
djangorestframework==3.14.0
lxml==4.9.2
mysqlclient==2.1.1
pytz==2022.7.1
requests==2.28.2
selenium==4.7.2
webdriver-manager==3.8.5